Struct teensy4_pins::Config

pub struct Config { /* private fields */ }

A configuration capable of compile-time, const configuration:

use imxrt_iomuxc::{Config, SlewRate, Hysteresis};

const UART_TX_CONFIG: Config = Config::zero()
    .set_slew_rate(SlewRate::Fast)
    .set_hysteresis(Hysteresis::Enabled);

Use configure() to set configurations to pads.

Implementations

impl Config

pub const fn zero() -> Config

Create a Config that will zero any unspecified field

This may not represent any registerโs reset value. A config thatโs created using zero() will have the effect of writing all fields to the register. Those that are not set explicitly set are written as zero.

use imxrt_iomuxc::{
    Config, configure, SlewRate,
    Hysteresis, PullKeeper, DriveStrength
};

// Set a configuration
configure(
    &mut gpio_ad_b0_13,
    Config::zero()
        .set_slew_rate(SlewRate::Fast)
        .set_drive_strength(DriveStrength::R0_7)
);
// DriveStrength::R0_7 as u32 | SlewRate::Fast as u32

// Completely change that configuration
configure(
    &mut gpio_ad_b0_13,
    Config::zero()
        .set_hysteresis(Hysteresis::Enabled)
        .set_pull_keeper(Some(PullKeeper::Pullup22k))
);
// Hysteresis::Enabled as u32 | PullKeeper::Pullup22k as u32
//
// Notice that SlewRate and DriveStrength were set to zero.

pub const fn modify() -> Config

Create a Config that will only modify the specified fields

Any field that is is not specified in the configuration will not be touched.

use imxrt_iomuxc::{Config, configure, SlewRate, DriveStrength, Hysteresis};

// Assume a starting value in the register...
configure(
    &mut gpio_ad_b0_13,
    Config::zero()
        .set_slew_rate(SlewRate::Fast)
        .set_drive_strength(DriveStrength::R0_7)
);
// DriveStrength::R0_7 as u32 | SlewRate::Fast as u32

// Now change the slew rate, and add hysteresis
configure(
    &mut gpio_ad_b0_13,
    Config::modify()
        .set_slew_rate(SlewRate::Slow)
        .set_hysteresis(Hysteresis::Enabled)
);
// DriveStrength::R0_7 as u32 | Hysteresis::Enabled as u32
//
// Notice that the DriveStrength field is preserved.

pub const fn is_zero(&self) -> bool

Returns true if this Config was created using zero(), meaning that it will zero any unspecified fields. If false, this config was created using modify(), which will not touch unspecified fields.

use imxrt_iomuxc::Config;

assert!(Config::zero().is_zero());
assert!(!Config::modify().is_zero());

pub const fn set_hysteresis(self, hys: Hysteresis) -> Config

Set the hysteresis bit

pub const fn set_pull_keeper(self, pk: Option<PullKeeper>) -> Config

Set the pull up / pull down / keeper configuration.

A None value disables the pull / keeper function.

pub const fn set_pullupdown(self, pud: PullUpDown) -> Config

๐Deprecated since 0.2.0: Use PullKeeper and Config::set_pull_keeper

Set the pull-up / pull-down value

pub const fn set_pull_keep_select(self, pke: PullKeepSelect) -> Config

๐Deprecated since 0.2.0: Use PullKeeper and Config::set_pull_keeper

Set the the pull-up / pull-down or keeper selection bit

pub const fn set_pull_keep(self, pk: PullKeep) -> Config

๐Deprecated since 0.2.0: Use PullKeeper and Config::set_pull_keeper

Set the flag that enables the keeper or pull-up / pull-down configuration

pub const fn set_open_drain(self, od: OpenDrain) -> Config

Set the open drain value

pub const fn set_speed(self, speed: Speed) -> Config

Set the pin speed

pub const fn set_drive_strength(self, dse: DriveStrength) -> Config

Set the drive strength

pub const fn set_slew_rate(self, sre: SlewRate) -> Config

Set the slew rate