Struct teensy4_pins::t40::Pins

pub struct Pins {

    pub p0: P0,
    pub p1: P1,
    pub p2: P2,
    pub p3: P3,
    pub p4: P4,
    pub p5: P5,
    pub p6: P6,
    pub p7: P7,
    pub p8: P8,
    pub p9: P9,
    pub p10: P10,
    pub p11: P11,
    pub p12: P12,
    pub p13: P13,
    pub p14: P14,
    pub p15: P15,
    pub p16: P16,
    pub p17: P17,
    pub p18: P18,
    pub p19: P19,
    pub p20: P20,
    pub p21: P21,
    pub p22: P22,
    pub p23: P23,
    pub p24: P24,
    pub p25: P25,
    pub p26: P26,
    pub p27: P27,
    pub p28: P28,
    pub p29: P29,
    pub p30: P30,
    pub p31: P31,
    pub p32: P32,
    pub p33: P33,
    pub p34: P34,
    pub p35: P35,
    pub p36: P36,
    pub p37: P37,
    pub p38: P38,
    pub p39: P39,

}

Teensy 4.0 pins

See from_pads to safely constrain the processor’s pads, and acquire Teensy 4.0 pins. Or, use new to unsafely create pins.

Fields
p0: P0

Pin 0

p1: P1

Pin 1

p2: P2

Pin 2

p3: P3

Pin 3

p4: P4

Pin 4

p5: P5

Pin 5

p6: P6

Pin 6

p7: P7

Pin 7

p8: P8

Pin 8

p9: P9

Pin 9

p10: P10

Pin 10

p11: P11

Pin 11

p12: P12

Pin 12

p13: P13

Pin 13

p14: P14

Pin 14

p15: P15

Pin 15

p16: P16

Pin 16

p17: P17

Pin 17

p18: P18

Pin 18

p19: P19

Pin 19

p20: P20

Pin 20

p21: P21

Pin 21

p22: P22

Pin 22

p23: P23

Pin 23

p24: P24

Pin 24

p25: P25

Pin 25

p26: P26

Pin 26

p27: P27

Pin 27

p28: P28

Pin 28

p29: P29

Pin 29

p30: P30

Pin 30

p31: P31

Pin 31

p32: P32

Pin 32

p33: P33

Pin 33

p34: P34

Pin 34

p35: P35

Pin 35

p36: P36

Pin 36

p37: P37

Pin 37

p38: P38

Pin 38

p39: P39

Pin 39

Implementations

impl Pins

pub const unsafe fn new() -> Self

Create an instance of Pins when you do not have a handle to the processor pads

Safety

Caller must ensure that the pins are not aliased elsewhere in the program. This could include

an existing handle to the imxrt-iomuxc pads,

another instance of Pins that was safely acquired using from_pads.

pub fn erase(self) -> ErasedPins

Erase the types of all pins