Module teensy4_pins::tmm

Teensy MicroMod specific APIs

Use from_pads to constrain the processor pads into the pins available on the Teensy MicroMod. If you cannot safely acquire all processor pads, use the unsafe Pins::new method to generate pins.

Pin	Pad ID	Alt0	Alt1	Alt2	Alt3	Alt4	Alt5	Alt6	Alt7	Alt8	Alt9

34	GPIO_SD_B0_03	USDHC1_DATA1	FLEXPWM1_PWMB01	LPUART8_RTS_B	XBAR1_INOUT07	LPSPI1_SDI	GPIO3_IO15	—	—	ENET2_RDATA00	SEMC_CLK6

35	GPIO_SD_B0_02	USDHC1_DATA0	FLEXPWM1_PWMA01	LPUART8_CTS_B	XBAR1_INOUT06	LPSPI1_SDO	GPIO3_IO14	—	—	ENET2_RX_ER	SEMC_CLK5

36	GPIO_SD_B0_01	USDHC1_CLK	FLEXPWM1_PWMB00	LPI2C3_SDA	XBAR1_INOUT05	LPSPI1_PCS0	GPIO3_IO13	FLEXSPIB_SS1_B	—	ENET2_TX_CLK	ENET2_REF_CLK2

37	GPIO_SD_B0_00	USDHC1_CMD	FLEXPWM1_PWMA00	LPI2C3_SCL	XBAR1_INOUT04	LPSPI1_SCK	GPIO3_IO12	FLEXSPIA_SS1_B	—	ENET2_TX_EN	SEMC_DQS4

38	GPIO_SD_B0_04	USDHC1_DATA2	FLEXPWM1_PWMA02	LPUART8_TX	XBAR1_INOUT08	FLEXSPIB_SS0_B	GPIO3_IO16	CCM_CLKO1	—	ENET2_RDATA01	—

39	GPIO_SD_B0_05	USDHC1_DATA3	FLEXPWM1_PWMB02	LPUART8_RX	XBAR1_INOUT09	FLEXSPIB_DQS	GPIO3_IO17	CCM_CLKO2	—	ENET2_RX_EN	—

40	GPIO_B0_04	LCD_DATA00	QTIMER2_TIMER1	LPI2C2_SCL	ARM_TRACE0	FLEXIO2_FLEXIO04	GPIO2_IO04	SRC_BOOT_CFG00	—	ENET2_TDATA03	—

41	GPIO_B0_05	LCD_DATA01	QTIMER2_TIMER2	LPI2C2_SDA	ARM_TRACE1	FLEXIO2_FLEXIO05	GPIO2_IO05	SRC_BOOT_CFG01	—	ENET2_TDATA02	—

42	GPIO_B0_06	LCD_DATA02	QTIMER3_TIMER0	FLEXPWM2_PWMA00	ARM_TRACE2	FLEXIO2_FLEXIO06	GPIO2_IO06	SRC_BOOT_CFG02	—	ENET2_RX_CLK	—

43	GPIO_B0_07	LCD_DATA03	QTIMER3_TIMER1	FLEXPWM2_PWMB00	ARM_TRACE3	FLEXIO2_FLEXIO07	GPIO2_IO07	SRC_BOOT_CFG03	—	ENET2_TX_ER	—

44	GPIO_B0_08	LCD_DATA04	QTIMER3_TIMER2	FLEXPWM2_PWMA01	LPUART3_TX	FLEXIO2_FLEXIO08	GPIO2_IO08	SRC_BOOT_CFG04	—	ENET2_RDATA03	—

45	GPIO_B0_09	LCD_DATA05	QTIMER4_TIMER0	FLEXPWM2_PWMB01	LPUART3_RX	FLEXIO2_FLEXIO09	GPIO2_IO09	SRC_BOOT_CFG05	—	ENET2_RDATA02	—

Re-exports

pub use crate::common::*;

Structs

Pins

Teensy MicroMod pins

Functions

from_pads

Constrain the processor pads to the Teensy MicroMod pins

Type Aliases

ErasedPins

Type-erased Teensy MicroMod pins

P34

Pin 34 (MicroMod)

P35

Pin 35 (MicroMod)

P36

Pin 36 (MicroMod)

P37

Pin 37 (MicroMod)

P38

Pin 38 (MicroMod)

P39

Pin 39 (MicroMod)

P40

Pin 40 (MicroMod)

P41

Pin 41 (MicroMod)

P42

Pin 42 (MicroMod)

P43

Pin 43 (MicroMod)

P44

Pin 44 (MicroMod)

P45

Pin 45 (MicroMod)